+++ python/dev-requirements.txt
@@ -4,4 +4,5 @@ pytest ==8.3.2
 pytest-asyncio ==0.23.8
 pytest-xdist ==3.6.1
 pytest-recording ==0.13.2
+opentelemetry-sdk ==1.26.0
 uvloop ==0.19.0; platform_system != 'Windows'
