+++ python/instrumentation/openinference-instrumentation-guardrails/tests/test_instrumentor.py
@@ -14,9 +14,7 @@ from guardrails.validator_base import (  # type: ignore[import-untyped]
 )
 from openinference.instrumentation import OITracer
 from openinference.instrumentation.guardrails import GuardrailsInstrumentor
-from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from opentelemetry.sdk.trace.export.in_memory_span_exporter import InMemorySpanExporter
 from pydash.strings import words as _words
 
@@ -49,19 +47,6 @@ class TwoWords(Validator):  # type: ignore[misc]
                 fix_value=self._get_fix_value(str(value)),
             )
         return PassResult()
-
-
-@pytest.fixture()
-def in_memory_span_exporter() -> InMemorySpanExporter:
-    return InMemorySpanExporter()
-
-
-@pytest.fixture()
-def tracer_provider(in_memory_span_exporter: InMemorySpanExporter) -> TracerProvider:
-    resource = Resource(attributes={})
-    tracer_provider = TracerProvider(resource=resource)
-    tracer_provider.add_span_processor(SimpleSpanProcessor(in_memory_span_exporter))
-    return tracer_provider
 
 
 @pytest.fixture()
