+++ python/instrumentation/openinference-instrumentation-langchain/tests/conftest.py
@@ -1,12 +1,22 @@
+import asyncio
+import logging
 import random
+from asyncio.events import BaseDefaultEventLoopPolicy
 from itertools import count
-from typing import Iterator
+from typing import Any, Dict, Iterator
 
 import pytest
 from openinference.instrumentation.langchain import LangChainInstrumentor
-from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk import trace as trace_sdk
 from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from opentelemetry.sdk.trace.export.in_memory_span_exporter import InMemorySpanExporter
+from opentelemetry.trace import TracerProvider
+
+for name, logger in logging.root.manager.loggerDict.items():
+    if name.startswith("openinference.") and isinstance(logger, logging.Logger):
+        logger.setLevel(logging.DEBUG)
+        logger.handlers.clear()
+        logger.addHandler(logging.StreamHandler())
 
 
 @pytest.fixture
@@ -15,10 +25,31 @@ def in_memory_span_exporter() -> InMemorySpanExporter:
 
 
 @pytest.fixture
-def tracer_provider(in_memory_span_exporter: InMemorySpanExporter) -> TracerProvider:
-    tracer_provider = TracerProvider()
+def tracer_provider(
+    in_memory_span_exporter: InMemorySpanExporter,
+) -> TracerProvider:
+    tracer_provider = trace_sdk.TracerProvider()
     tracer_provider.add_span_processor(SimpleSpanProcessor(in_memory_span_exporter))
     return tracer_provider
+
+
+@pytest.fixture(scope="session")
+def event_loop_policy() -> BaseDefaultEventLoopPolicy:
+    try:
+        import uvloop
+    except ImportError:
+        return asyncio.DefaultEventLoopPolicy()
+    return uvloop.EventLoopPolicy()
+
+
+@pytest.fixture
+def vcr_config() -> Dict[str, Any]:
+    return dict(
+        before_record_request=lambda _: _.headers.clear() or _,
+        before_record_response=lambda _: {**_, "headers": {}},
+        decode_compressed_response=True,
+        ignore_localhost=True,
+    )
 
 
 @pytest.fixture(autouse=True)
@@ -29,7 +60,6 @@ def instrument(
     LangChainInstrumentor().instrument(tracer_provider=tracer_provider)
     yield
     LangChainInstrumentor().uninstrument()
-    in_memory_span_exporter.clear()
 
 
 @pytest.fixture(autouse=True)
