+++ python/openinference-semantic-conventions/tests/test_version.py
